feat: Steps to switch from Spring Boot to Spring Core
Basically, we need to do these things. One, importing dependencies from
the core library so we could get the application started; Second, change
the "smart" annotation to explicit annotations underneath; Third, add
the code for closing the application as when we're using Spring Boot, it
would do them for us.